+++ handlers/salesforce-disaster-recovery/riff-raff.yaml
@@ -21,6 +21,6 @@ deployments:
       prefixStack: false
       functionNames:
         - save-salesforce-query-result-to-s3-
-        - update-chunk-zuora-accounts-
+        - update-zuora-accounts-
         - divide-into-chunks-
     dependencies: [salesforce-disaster-recovery-cloudformation]
